+++ src/main/java/sk/freemap/gpxAnimator/CommandLineConfigurationFactory.java
@@ -14,6 +14,8 @@
  */
 package sk.freemap.gpxAnimator;
 
+import sk.freemap.gpxAnimator.ui.TrackIcon;
+
 import java.awt.Color;
 import java.awt.GraphicsEnvironment;
 import java.io.File;
@@ -35,7 +37,7 @@ public final class CommandLineConfigurationFactory {
 
     private final List<Float> lineWidthList = new ArrayList<Float>();
 
-    private final List<Boolean> iconEnableList = new ArrayList<Boolean>();
+    private final List<TrackIcon> trackIconList = new ArrayList<TrackIcon>();
 
     private final boolean gui;
 
@@ -108,8 +110,8 @@ public final class CommandLineConfigurationFactory {
                         case INPUT:
                             inputGpxList.add(args[++i]);
                             break;
-                        case ICON_ENABLE:
-                            iconEnableList.add(Boolean.valueOf(args[++i]));
+                        case TRACK_ICON:
+                            trackIconList.add(new TrackIcon(args[++i]));
                             break;
                         case KEEP_IDLE:
                             cfg.skipIdle(false);
@@ -207,7 +209,7 @@ public final class CommandLineConfigurationFactory {
             tcb.label(i < labelList.size() ? labelList.get(i) : "");
             tcb.timeOffset(i < timeOffsetList.size() ? timeOffsetList.get(i) : Long.valueOf(0));
             tcb.forcedPointInterval(i < forcedPointIntervalList.size() ? forcedPointIntervalList.get(i) : null);
-            tcb.enableIcon(iconEnableList.get(i));
+            tcb.trackIcon(trackIconList.get(i));
 
             cfg.addTrackConfiguration(tcb.build());
         }