+++ src/main/java/sk/freemap/gpxAnimator/Option.java
@@ -44,7 +44,7 @@ public enum Option {
     TILE_CACHE_PATH("tile-cache-path", "path to a directory to use for caching map tiles"),
     TILE_CACHE_TIME_LIMIT("tile-cache-timeout", "time a cached map tile is valid"),
     HELP("help", "this help"),
-    ICON_ENABLE("ICON ENABLE", "enables a bicycle icon that replaces a standard circle of the current position.");
+    TRACK_ICON("track-icon", "enables an icon that replaces the standard circle of the current position (only one valid value: Bicycle)");
 
     private static java.util.Map<String, Option> map = new HashMap<String, Option>();